+++ dwds/lib/src/loaders/legacy.dart
@@ -11,7 +11,38 @@ class LegacyStrategy extends LoadStrategy {
   @override
   final ReloadConfiguration reloadConfiguration;
 
-  LegacyStrategy(this.reloadConfiguration);
+  /// Returns the module for the corresponding server path.
+  ///
+  /// For example:
+  ///
+  /// /packages/path/path.ddc.js -> packages/path/path
+  ///
+  final String Function(String sourcePath) _moduleForServerPath;
+
+  /// Returns the server path for the provided module.
+  ///
+  /// For example:
+  ///
+  ///   web/main -> main.ddc.js
+  ///
+  final String Function(String module) _serverPathForModule;
+
+  /// Returns the server path for the app uri.
+  ///
+  /// For example:
+  ///
+  ///   org-dartlang-app://web/main.dart -> main.dart
+  ///
+  /// Will return `null` if the provided uri is not
+  /// an app URI.
+  final String Function(String appUri) _serverPathForAppUri;
+
+  LegacyStrategy(
+    this.reloadConfiguration,
+    this._moduleForServerPath,
+    this._serverPathForModule,
+    this._serverPathForAppUri,
+  );
 
   @override
   Handler get handler => (request) => null;
@@ -35,4 +66,14 @@ class LegacyStrategy extends LoadStrategy {
   @override
   String loadClientSnippet(String clientScript) =>
       'window.\$dartLoader.forceLoadModule("$clientScript");\n';
+
+  @override
+  String moduleForServerPath(String serverPath) =>
+      _moduleForServerPath(serverPath);
+
+  @override
+  String serverPathForModule(String module) => _serverPathForModule(module);
+
+  @override
+  String serverPathForAppUri(String appUri) => _serverPathForAppUri(appUri);
 }