+++ dwds/test/handlers/injected_handler_test.dart
@@ -39,6 +39,15 @@ class FakeStrategy implements LoadStrategy {
 
   @override
   String loadClientSnippet(String clientScript) => 'dummy-load-client-snippet';
+
+  @override
+  String moduleForServerPath(String serverPath) => null;
+
+  @override
+  String serverPathForModule(String module) => null;
+
+  @override
+  String serverPathForAppUri(String appUri) => null;
 }
 
 void main() {